% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\name{calculate_segregation}
\alias{calculate_segregation}
\title{Calculate segregation}
\usage{
calculate_segregation(macro_geometry_sf, sub_geometry_sf, majority_pop,
  minority_pop, method = c("dissimilarity_index", "gini", "entropy",
  "interaction"))
}
\arguments{
\item{macro_geometry_sf}{An `sf` object. Contains simple features for
macro-geometries (eg cities)}

\item{sub_geometry_sf}{An `sf` object. Contains simple features for
sub-geometries (eg tracts or neighborhoods) and demographic data.}

\item{majority_pop}{Character. The name of the column with the majority
race/ ethnicity county.}

\item{minority_pop}{Character. The name of the column with the minority
race/ ethnicity county.}

\item{method}{Character. The method to use to calculate gentrification.
One of "dissimilarity_index", "gini", "entropy", or "interaction".
Currently only one at a time is supported.}
}
\value{
`macro_geometry_sf` with one more column containing the segregation
  measure selected.
}
\description{
Calculate segregation for macro geometries. Method definitions from the
\href{https://www.census.gov/hhes/www/housing/resseg/pdf/app_b.pdf}{US
Census}.
}
