% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_matching.R
\name{get_sub_geometries}
\alias{get_sub_geometries}
\title{Get Sub-Geometries}
\usage{
get_sub_geometries(macro_geometry_sf, sub_geometry_sf,
  wt_threshold = 0.01)
}
\arguments{
\item{macro_geometry_sf}{An `sf` object for one macro geometry}

\item{sub_geometry_sf}{An `sf` object for sub-geometries that might be in the
macro geometry}

\item{wt_threshold}{The threshold of intersecting area to filter out.
Defaults to 0.01.}
}
\value{
An tidy `sf` for tracts that intersect the neighborhood
}
\description{
Get the sub-geometries for a macro geometry
}
